Add server and app files
